fix(i18n): add i18n module to provide i18n export, update all '@/main' imports causing cyclic import errors to use module import
