+++ src/models/Language.ts
@@ -4,7 +4,7 @@
 import detectBrowserLanguage from 'detect-browser-language';
 import _ from 'lodash';
 import CCUModel from '@/models/base';
-import { i18n } from '@/main';
+import { i18n } from '@/modules/i18n';
 
 interface LanguageTranslationResponse {
   text: string;
