+++ src/modules/i18n.ts
@@ -1,0 +1,13 @@
+import { createI18n } from 'vue-i18n';
+
+const getI18n = (messages = {}) => {
+  return createI18n({
+    legacy: false,
+    formatFallbackMessages: true,
+    silentFallbackWarn: false,
+    locale: 'en',
+    messages,
+  });
+};
+
+export const i18n = getI18n();
